fix Рефакторинг после ревью
